requirements.txt
numpy
pandas
tensorflow-cpu
sklearn